:sparkles: (Thread): Created isolated model
Since I ran into difficulties with relating a model twice in separate properties to another model, I decided to comment them out for now. The main driving parts of the model are there, like the title, body, upvotes, and relation to a module. Relating the thread to an author (User), watchers (User[]), and comments were not successful so far.