feat(LP): created base schema and model structure (#493)
* feat(LP): created base schema and model structure

* fix(date): Changed date returned by scalar for more accurate sorting

* fix(LP): created base query and mutation
The query retrieves through the user's Plan ID. The mutation expects the planID and the path structure directly inputted

* fix(LP): mutation creates enrollments into each module passed in

* chore(prisma): Bumped version number

* chore(ts): Added type inference rule

* fix(LP): Modified schema structure
Now a single LP model is assigned to a single plan of study, creating a 1 to 1 relationship between the LP model and the user. However, now the LP model stores an array of paths, making the functionality nearly the same as users can still have one to many relations with paths, its just one level deeper in the composite type.

* feat(LP): Added basic safeguards to creation
As the input object accepts either a single path object or a list of path objects, I added a statement that checks if either of the two is passed in before continuing with the operations

* feat(LP): Started implementing union path input functionality
Mainly focused on getting the single path input working. Still need to match and create the module enrollments before we proceed to validating the path data

* feat(LP): added LP status enum to schema
The enum can be either Draft or Live. Once users update their paths to be live, they get enrolled into the modules. While the path is in draft, they aren't enrolled in any of the modules.

* feat(LP): removed enrollment creation from service
This operation will be handled by the update service rather then the initial creation service

* feat(LP): added hours satisfied and learning outcomes to path model
Hours satisfied tracks the number of hours that the current enrollment in the path satisfies. Learning outcome of the path defines all the learning outcomes of the modules being taken

* feat(LP): added name field to sections and collections
Since we are using fake data for the sections and collections, we need something visual to create UI layouts with. Eventually only the ID will be required, and the service will retrieve the appropriate data and concatenate it with the ID

* feat(LP): manually extended course and module types
Since we want to get the appropriate data from GQL, we want to allow the selection of subfields that match the composite type

* fix(LP): resolve concatenation issue with query
Previously, the query returned a single section with a single collection and mapped all the modules into that one element array. Now the query keeps the existing composite structure and fetches the data to concatenate into the module object.

* fix(LP): resolved fields not receiving values on creation
I forgot to add the learning outcomes and hoursSatisfied fields to the create operation. Now we also store collection and section names in the composite types until we get the program structure updated

* feat(LP): added optional filter for LPs by pathID

* feat(LP): implemented mutation to update paths
The mutation takes in the inputted course structure and replaces the existing one in the array while keeping the other records untouched

* fix(LP): added required import statement

* feat(LP): created mutation to delete a single path
User can now delete individual paths from an LP model

* feat(LP): created mutation to create a single path
User can now create individual paths attached to an LP model

* fix(LP): resolved schema update conflicts

* refactor(user): added mutation to handle create operation

* fix(sections): argument in GQL and the API were mismatching

* fix(schema): turned collection to module relation to many-to-many
Since collections are just grouped modules, a module could be part of many collections and collections can hold a number of modules.

* fix(progress): resumption needed to change according to FE

* feat(lp): updated includes

* refactor(lp): updated path schema to match redevelopment

* feat(module): added query to get data by learning Path

* feat(module): additional fields added as required by FE
Modules have an optional direct relation to an instructor profile to connect instructors with their modules. Furthermore, I added module prefixes (ENMA, CS, DS, etc), and module numbers

* feat(course): additional fields added as needed
I added course prefixes (ENMA, CS, DS, etc), and course numbers, so not everything is stored in the course name field. The name field should be reserved for the user friendly title of the course.

* refactor(course): creation of many courses at once is supported

* fix(type): addressed type errors as a result of schema changes

* feat(modules): created direct resolver for module flow and return type to go with it
Return type to be added to the resolver once all the edge cases are figured out

* refactor(LP): modified the selectable fields for sections

---------

Co-authored-by: Dániel B. Papp <chef-danny-d@users.noreply.github.com>